Add script for Git commit message generation
Add a script that generates a commit title and message by passing the
diff into an LLM and requesting it to summarize the diff.

Currently using Ollama to run a local model. Use of a different API (e.g.
OpenAI) is something that can be added later.

This will be adjusted to fit the format of Git's prepare-commit-msg hook.